Q version 57.1: fix 1 compiler warning
Updated:

 q/q_version.h: 57.1

 q/scrdit.c: Cast thisch for -ve test (picked up on OpenBSD system gcc 4.2.1)

+++ q/scrdit.c
@@ -1044,7 +1044,7 @@ process_pseudo_arg(scrbuf5 *Curr)
   if (gwthr)
   {
     qreg = -1;                     /* Default response: no such macro */
-    if (thisch < 0)
+    if ((int16_t)thisch < 0)
       ;                            /* Can't be a macro */
     else if (thisch == 0 && curmac >= 0)
       qreg = 0;                    /* ^N^@ legal in a macro */
